[PR #5744 @nhat-nguyen] Remove unused parameter in DoStackArgsOpt [Update submodule only]
`DoStackArgsOpt` takes a `topFunc` parameter but doesn't use it.

[Core Commit Message]

[MERGE #5744 @nhat-nguyen] Remove unused parameter in DoStackArgsOpt

Merge pull request #5744 from nhat-nguyen:stackargs

`DoStackArgsOpt` takes a `topFunc` parameter but doesn't use it.

It seems like the intention is to pass it to `IsStackArgsEnabled`, but `topFunc` is actually referenced inside `IsStackArgsEnabled` through a pointer in m_func, making passing in the argument unnecessary.