[MERGE #5758 @sethbrenith] Use different machine pool for OSX CI builds
Merge pull request #5758 from sethbrenith:user/sethb/osx-ci

This is the combination of several changes that have already gone into release/1.10-ci. It updates our CI configuration to build on a different set of Macs, and fixes unit tests that fail with the version of ICU installed on those machines.